+++ apps/chat/src/components/Chat/Publish/PublicationHandler/ReviewApplicationDialog/ReviewApplicationDialogView.tsx
@@ -15,8 +15,8 @@ import { useAppSelector } from '@/src/store/hooks';
 import { ApplicationSelectors } from '@/src/store/selectors';
 
 import { PublicationControls } from '@/src/components/Chat/Publish/PublicationChatControls';
-import { ReviewCodeAppSection } from '@/src/components/Chat/Publish/ReviewApplicationDialog/ReviewCodeAppSection';
-import { ReviewQuickAppSection } from '@/src/components/Chat/Publish/ReviewApplicationDialog/ReviewQuickAppSection';
+import { ReviewCodeAppSection } from '@/src/components/Chat/Publish/PublicationHandler/ReviewApplicationDialog/ReviewCodeAppSection';
+import { ReviewQuickAppSection } from '@/src/components/Chat/Publish/PublicationHandler/ReviewApplicationDialog/ReviewQuickAppSection';
 import { ModelIcon } from '@/src/components/Chatbar/ModelIcon';
 import { withRenderWhen } from '@/src/components/Common/RenderWhen';
 import { ApplicationTopic } from '@/src/components/Marketplace/ApplicationTopic';
