+++ apps/chat/src/components/Chat/Publish/PublicationHandler/ReviewApplicationDialog/ReviewQuickAppSection.tsx
